pyproject.toml
[tool.poetry]
name = "spotify-etl"
version = "0.1.0"
description = "Spotify ETL Project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "spotify_etl"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
